Move `Packet.h` (without writeName) into `Core/Data`, implement `NetworkPacket` and `HostPacket`
Old compilers require constructor for inherited classes
